Use hpctoolkit persistent id

+++ src/redundancy.cpp
@@ -119,7 +119,7 @@ void record_temporal_trace(PCPairs &pc_pairs, PCAccessCount &pc_access_count, u3
   record_data.num_views = num_views;
 }
 
-void show_temporal_trace(u32 thread_id, u64 kernel_id, u64 total_red_count, u64 total_count,
+void show_temporal_trace(u32 thread_id, i32 kernel_id, u64 total_red_count, u64 total_count,
                          TemporalStatistics &temporal_stats, bool is_read, bool is_thread) {
   using std::endl;
   using std::get;
@@ -258,7 +258,7 @@ void record_spatial_trace(SpatialTrace &spatial_trace, PCAccessCount &pc_access_
   record_data.num_views = num_views;
 }
 
-void show_spatial_trace(u32 thread_id, u64 kernel_id, u64 total_red_count, u64 total_count,
+void show_spatial_trace(u32 thread_id, i32 kernel_id, u64 total_red_count, u64 total_count,
                         SpatialStatistics &spatial_stats, bool is_read, bool is_thread) {
   using std::endl;
   using std::get;
